Clean up LocationTickPhase
